+++ ButtonCore/src/main/resources/plugin.yml
@@ -9,3 +9,5 @@ commands:
     description: Schedules a restart for a given time.
   primerestart:
     description: Restarts the server as soon as nobody is online.
+  randomtp:
+    description: teleport player to random location within world border. Every five players teleport to the same general area, and then a new general area is randomly selected for the next five players.
