--- ChunkArchive/.gitignore
@@ -1,2 +1,0 @@
-/target/
-upload_key
