--- ChunkArchive/README.md
@@ -1,2 +1,0 @@
-# ChunkArchive
-Archives chunks as they're populated
