--- ChunkArchive/src/main/resources/plugin.yml
@@ -1,3 +1,0 @@
-name: ChunkArchive
-main: buttondevteam.ChunkArchive.Main
-version: 4.0
