--- RandomTeleport/README.md
@@ -1,2 +1,0 @@
-# RandomTeleport
-Not randomly teleports players to random places. (RTP) /randomtp
