--- RandomTeleport/plugin.yml
@@ -1,6 +1,0 @@
-main: randomTP.Main
-version: 1.0.0
-name: RandomTP
-commands:
-  randomtp:
-    description: teleport player to random location within world border. Every five players teleport to the same general area, and then a new general area is randomly selected for the next five players.
